requirements.txt
py-tgcalls==0.8.1rc1
git+https://github.com/pyrogram/pyrogram
tgcrypto
youtube-search-python
pytube
spotipy
python-dotenv
youtube-dl